requirements/base.txt
channels
channels-redis
Django
django-environ
django-extensions
djangorestframework
psycopg2-binary
celery[redis]
Pillow
